pyproject.toml
[tool.poetry]
name = "eikumi"
version = "0.1.0"
description = "A moderation discord bot for the IEC."
authors = ["Ohin \"Kazani\" Taylor <kazani@kazani.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
rich = "^13.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
